more refactoring and code cleanup
