+++ front-end/src/components/userRecipes.js
@@ -3,64 +3,11 @@ import { connect } from "react-redux";
 import { useHistory } from "react-router-dom";
 import { recipeFormValues, addRecipe } from "../actions/index";
 import { Button, TextField } from "@material-ui/core";
-import { makeStyles } from "@material-ui/core/styles";
+import useStyles from '../styles/styles'
 import { axiosWithAuth } from "../utils/axiosWithAuth";
 import RecipeCard from "./recipeCard";
+import { initialFormState } from '../initialStates/initialStates'
 
-const initialFormState = {
-  user_id: "",
-  recipe_name: "",
-  recipe_source: "",
-  image_source: "",
-  category: "",
-  recipe_description: "",
-  ingredients: [],
-  instructions: [],
-};
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    maxWidth: "50vw",
-    alignSelf: "left",
-    padding: "3vh 5vw",
-    backgroundColor: "#222",
-    opacity: "0.9",
-    "& > * + *": {
-      marginTop: theme.spacing(2),
-    },
-  },
-  input: {
-    color: "white",
-    backgroundColor: "#AAA",
-    boxShadow: "0 0 2vh #333"
-  },
-  button: {
-    color: "black",
-    backgroundColor: "#FFF",
-    margin: "2vh 2vw",
-    maxWidth: "10rem",
-  },
-  outerDiv: {
-    marginTop: '10vh',
-    display: 'flex',
-    justifyContent: 'center',
-    flexDirection: 'row'
-  },
-  h4: {
-  marginTop: "3vh"
-  },
-  label: {
-    margin: "3vh 0", 
-    color: 'white'
-  },
-  select: {
-    backgroundColor: "#444",
-    color: "#FFF",
-    boxShadow: "0 0 2vh #333",
-    margin: '2vh 0'
-  },
-
-}));
 
 const UserRecipes = (props) => {
   const classes = useStyles();
@@ -228,9 +175,9 @@ const UserRecipes = (props) => {
           <div  
             className="flexible-stretch-boxes d-flex justify-content-center flex-column"
             style={{
-            padding: '2vh auto',
+            padding: '2vh 4vw',
             color: 'white',
-            fontSize: '3vh',
+            fontSize: '3.5vh',
             margin: "2vh auto",
             }}>
             <label
