+++ front-end/src/initialStates/initialStates.js
@@ -1,0 +1,11 @@
+export const initialFormState = {
+    user_id: "",
+    recipe_name: "",
+    recipe_source: "",
+    image_source: "",
+    category: "",
+    recipe_description: "",
+    ingredients: [],
+    instructions: [],
+  };
+
